backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aimemo-backend"
version = "0.1.0"
description = "AiMemo Tornado backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "tornado>=6.4",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "passlib[bcrypt]>=1.7",
  "pyjwt>=2.9",
  "structlog>=24.1.0",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["ruff>=0.4", "black>=24.4"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
